refactor: [sync program] reorganize functions path
