+++ src/pages/Synchronization/Programs/ProgramGlobalSettings.js
@@ -10,8 +10,8 @@ import {
 } from '../../../constants/program-settings'
 import { SelectSettings } from '../../../components/inputs'
 import SettingsTable from '../../../components/settings-table/settings-table'
-import { parseValueByType } from '../../../modules/programs/parseValueBySettingType'
-import { populateProgramObject } from '../../../modules/programs/populateProgramObject'
+import { parseValueByType } from './parseValueBySettingType'
+import { populateProgramObject } from './populateProgramObject'
 import inputStyles from '../../../styles/Input.module.css'
 
 const CODE = 'settingDownload'
