+++ src/pages/Synchronization/Programs/SpecificTableAction.js
@@ -5,7 +5,7 @@ import TableActions from '../../../components/table-actions'
 import DialogDelete from '../../../components/dialog/dialog-delete'
 import DialogSpecificSetting from './DialogSpecificSetting'
 import { isProgramWithRegistration } from './helper'
-import { parseValueByType } from '../../../modules/programs/parseValueBySettingType'
+import { parseValueByType } from './parseValueBySettingType'
 import {
     removeSettingsFromList,
     updateSettingsList,
